requirements.txt
eventlet==0.25.1
# gevent==20.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
flask-nav==0.6
flask-paginate==0.5.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-SocketIO==4.2.1
Flask-RESTful==0.3.8
gunicorn==20.0.4
Jinja2==2.11.1
requests==2.23.0
SQLAlchemy==1.3.13
urllib3==1.25.8
virtualenv==20.0.13
Werkzeug==1.0.0
PyMySQL==0.9.3
gunicorn==20.0.4
xlwt==1.3.0
passlib==1.7.
Flask-HTTPAuth==4.0.0
Flask-Migrate==2.5.3
Flask-JWT==0.3.2
Flask-Caching==1.8.0
python-dateutil==2.8.2
tzlocal==4.2
